Added two new notebooks
Added a notebook that explores the Nyquist criterion, zero filling,
spectral clipping, and phase errors.

Also added a notebook that does some fun stuff with spectral smoothing,
which is an interesting application of Convolution and FT.

These still need a lot of work in terms of the additional text for each
section.