Removed plotting functions, import instead
Went through all of the notebooks, and removed the explicit definition
of the plotting functions. Instead, I do a %run on the plotting.py file
to import these functions from a global location. Made a couple of small
text modifications, and changed the colors of the FT plots in
plotting.py.