Abstracted Fourier Series Notebook
After feedback from Beth Stemmler, I abstracted a bunch of stuff from
the Fourier Series notebook. Specifically, all of the plotting stuff and
some of the calculations are hidden from the user. These things were
converted to functions that are now contained in the file
'fourier_series.py', which is imported at the beginning of the notebook.