Split Fourier Intro, Part 1
I split up the Fourier Intro notebook. The first part of this notebook
had a simple 4 point FT and the second part was FTs of sine and cosine
function. It seemed best to break these apart so that there are not so
many things going on in one notebook. I'll need to fiddle with the
second half of the notebook before committing it to the repo.