Split Fourier Intro, Part 2
Here I made a new Fourier Intro notebook that covers the FT of simple
sinusoidal functions and their complex mixtures and the inverse FT. I
also made a new plotting function file that will eventually supersede
the original.  The FT plotting function was modified to handle the
sorting of the frequency data automatically, so I didn't have to define
it in the text directly.